requirements.txt
fastapi
python-multipart
torch
apscheduler
uvicorn[standard]
openai-whisper
torch
psutil